Make data_browser file the init. This fixes circular imports. The old init imports were not needed in ScopeFoundry, but might be needed in other browsers that depend on the base browser.
